Fix to run with substrate 19.3 Java 11 support
When running with substrate the AtomosRuntimeModules runtime cannot be
used because ModuleReader is not functional in a substrate native image.
This fix allows the AtomosRuntimeSubstrate runtime to be used with Java
11 when running in a substrate native image. This still requires a
substrate_lib folder to be available with all the original bundle jars
so Atomos can load the bundle entries at runtime.