Add support for loading bundle entries from resources in native image
Up to now to run in a substrate native image the original bundle JARs
needed to exist in a substrate_lib folder and Atomos would discover the
JARs and assume the JARs found represented bundles included in the
native image.  With this new approach a resource is discovered that
lists the bundles included in the image and a mappeing of the resources
that belong to which bundles.  This eliminates the need to have the
original bundle JARs when running the native image.