Initial test with Felix Framework
This adds a Felix framework dependency for a
connect implementation.  The default still does
not use the Felix framework. To test with
Felix you need to change the following pom depenency
in atomos/atomos.framework/pom.xml

<dependency>
	<groupId>org.atomos.osgi.frameworks.equinox</groupId>
	<artifactId>atomos.osgi.framework</artifactId>
	<version>${atomos.version}</version>
</dependency>

Additional changes are needed to the jlink example
in atomos/atomos.tests/service.image/pom.xml
to switch.  Doing both cause issues currently that
need to be worked.