Remove usage of atomos bundles.index file
This example required the copying of resources
from bundles included in the image to the
source repository.  This was always meant to be
a temporary thing, but as of right now that
means we have some resources in the repo that
are not Apache licensed.  Removing them in
prep to move the code to Apache Felix where
we should not check in these resources

Long term plan is to have a maven plugin that
does this work automatically to avoid having
to copy such resources into the source repo