Use felix.cache.locking=false to avoid issues on windows
The testcase is an error condition causing Felix to
fail to release its cache lock.  This results in
failing to delete the temp dir on Windows for the
framework storage area. Disabling the cache lock
in felix to work around

Also, be sure to wait for stop on temp framework instance