FELIX-6227 - Generalize Atomos index to be used independent of substrate
The substrate support for Atomos has a concept of an
atomos/bundles.index resource which can be read at runtime to figure out
what bundles are available and what resources are available for bundle
entries.

This commit refactors that support to be general such that this resource
can be used for loading Atomos content from any environment where the
resource atomos/bundles.index can be found by the Atomos runtime. This
allows for other cases similar to substrate where the original bundle
JARs may no longer be around as artifacts at runtime.  For example, with
Android applications.