FELIX-6227 - Add configuration for path to Atomos index
Instead of using an atomos.load.index option with FIRST|IGNORE
this change adds an option atomos.index.path that allows
the specification of a path to the Atomos index resource.
If set to IGNORE then no Atomos index is loaded.
The path may begin with a '/', but is not required
to.  The value is the path to the index resource
and the parent path of the resource will be used
as the relative path for all the bundle indexes