Add example using the Spring loader
This example creates an executable jar using
the spring-boot-maven-plugin.  The result
is a single executable JAR with embedded
bundle JARs.  Atomos is able to work with
the jar URL handler the Spring loader provides
to load entries out of the embedded bundle JARs
with no need to extract the embedded JARs