Handle the persistence issues when changing to Java 8
When running with Java 11 or higher with the module path
the persistence for Atomos will contain at least 2 layers,
the empty and the boot layer. This causes issues if you the
launch the persistent configuration using Java 8 because
Atomos will only allow a single layer to exist in
that case.

This handles that case by printing a message and
continuing on as if there was no persistence