Add Android Example
This example requires JAR containing the bundles
and an Atomos index to be placed in the folder
atomos.examples/atomos.examples.android/app/libs/
For example, the on produced by the
atomos.examples.index project.

This project uses gradle to build because the
maven plugins to build Android are outdated
The Android SDK also needs to be setup to build
with gradle.  This project was created with
Android Studio.