Index lists all content and stores only duplicates
Change the Atomos bundles.index file to list
all entries of each bundle.  Only store duplicate
paths in the atomos/ index directory.  If a resource
entry cannot be found in the atomos/ index directory
then look in the root.  Only look in the root
if the lookup path is contained in the list
stored in the bundles.index file