FELIX-6313 - handle modules names that are different from BSN
JPMS modules may have a name that does not match the
Bundle-SymbolicName specified in the module bundle manifest.
OSGi require-bundle symantics closely match the symantics
of JMPS module requires.  The require-bundle wire in OSGi
is a natural fit when trying to map non-OSGi JPMS modules
as OSGi connect bundles in Atomos.  But this mapping depends
on the BSN and the module names to match.  When they don't
then the conversion of a module requires to a require-bundle
needs to account for the fact that the required module will
have a different BSN within the OSGi framework.

This change detects the mismatch and will use the correct
BSN for the require-bundle wire for the connect bundle.