Add Aries JAX-RS example
This example creates both an Atomos indexed uber JAR
and a substrate native image with a very simple
hello JAX-RS resource.  The resource is a DS
component that gets injected with another
OSGi service.

The maven plugin should have calculated
the necessary configuration files for reflection
but there seemed to be things missing. This example
has additional AgentConfig.json files that were gathered
by using the graal agent to discover the necessary
resources and reflect configs for substrate.