Stop using parallel stream
The Atomos maven plugin was using parallel streams in several places.
Unfortunately many of the consumers called in the parallel streams are
not thread safe.  This can lead to unpredictable behavior.  For example,
randomly the plugin would omit a few bundles to include in the
processing.

This change also sorts the atomos index file to be in bundle symbolic
name order.  This allows for a consistent install order for images that
use the Atomos index.