Set system packages correctly when modules from boot layer are loaded
The launcher needs to set the system packages to empty any time the boot
modules from the JVM are loaded as atomos content.  Otherwise the system
bundle will export duplicate packages that are exported by boot module
connect bundles.  This was not happening correctly for the case where
the old classpath mode was used on Java 11 because
org.apache.felix.atomos.runtime.AtomosLayer.isAddLayerSupported() was
returning false in this case.

The fix is to set the system packages to the empty string anytime the
AtomosRuntimeModules is used.